fix unittests for AxesToDataScale
